world/go_scripts.cpp: four handlers, two of which never ran
Four were live for 2.4.3; four more were WOTLK and later and left with the
file.

go_table_theka is a GameObjectScript bound in script_binding as type 0
(SCRIPTED_UNIT) against bind 7272 -- Theka the Martyr, a creature. The
gameobject path never found it. Quest 2936 carries SpecialFlags = 2 and no
required creature, object or item, so AreaExploredOrEventHappens is the only
thing that can complete it and nothing was calling it: the quest has been
uncompletable. The port binds it to gameobject 142715, the Tablet of Theka.
Its gossip half needed nothing -- SEND_GOSSIP_MENU(1653) passes an npc_text
id, the object's data3 is gossip menu 1053, and gossip_menu 1053 has
text_id 1653. The script was replacing the default with an identical one.

go_ethereum_prison switches on the summoned creature's faction to choose a
reputation spell and falls through to script_error_log. Five of the six
friendly prisoners carry faction template 35 in this world -- the generic
"Friendly" -- so five in six reach the error log and hand out nothing. The
port writes the mapping the switch was for, one branch per entry, and says
so in the file together with the alternative: five creature_template rows
and the C++ would have worked too.

What the file needed:

  random_script first= count=   one of THIRTEEN. The three-column form
                                would have been five rows and a tree.
  summon_at_target face=1       turned towards its summoner without
                                attacking it, which is what
                                pGo->Where().BearingTo(pPlayer->Where())
                                said.

  MaxOperands 8 -> 12. A position facet is four slots on its own, so any
  verb carrying one had four left, and summon_at_target used all four the
  moment it learnt a spawn mode, an order to attack and an order to face.
  The alternative was folding two named booleans into one flags column,
  which is precisely what this format exists to stop: attack=1 face=1 says
  what flags=3 does not. Sixteen bytes a step buys the naming back.

  Nothing at all for the tailoring check. `conditions` already says "has
  this skill at this value" (7) and "has, or has not, this spell" (17) and
  composes them with AND (-1); terminate_cond already reads a condition and
  buddy_flags 8 inverts it. Two new verbs would have been a second, worse
  copy of a table the core reads everywhere else.

Verified against a scratch copy of the live world.

+++ src/game/Scripting/mai/MaiLowering.cpp
@@ -137,7 +137,7 @@ namespace mai
             }
             if (Supplied(raw[i]) || !spec->params[slot].optional)
             {
-                out.given |= uint8(1u << slot);
+                out.given |= uint16(1u << slot);
             }
         }
 
@@ -151,7 +151,7 @@ namespace mai
                 // one, so the test here is not Supplied().
                 if (row.textId[i] >= 0)
                 {
-                    out.given |= uint8(1u << slot);
+                    out.given |= uint16(1u << slot);
                 }
             }
         }
@@ -162,7 +162,7 @@ namespace mai
             for (int i = 0; i < 4; ++i, ++slot)
             {
                 out.operands[slot].f = at[i];
-                out.given |= uint8(1u << slot);
+                out.given |= uint16(1u << slot);
             }
         }
